cloud/requirements.txt
google-cloud-compute
google-cloud-tpu
